requirement.txt
click==7.1.1
Flask==1.1.2
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
SQLAlchemy==1.3.16
Werkzeug==1.0.1
gensim==3.8.3
mysql-connector-python==8.0.17
Cython==0.29.14
